Moved stuff that needs no refactor
