+++ Legacy/src/gameObjects/card/board-card.ts
@@ -1,7 +1,7 @@
-import { EFFECT_ASSET_KEYS } from '../../assets/asset-keys'
+import { EFFECT_ASSET_KEYS } from '../../../../src/assets/asset-keys'
 import { setOutline } from '../../common/outline'
 import { BattleScene } from '../../scenes/battle-scene'
-import { BATTLE_STATES, TARGET_KEYS, TargetKeys, WARNING_KEYS } from '../../utils/keys'
+import { BATTLE_STATES, TARGET_KEYS, TargetKeys, WARNING_KEYS } from '../../../../src/utils/keys'
 import { CARD_SCALE, SUMMONING_SICK_CONFIGS, ZZZ_ANIMATION_POSITION } from '../../utils/visual-configs'
 import { Card } from './card'
 import { CardData } from './card-keys'
