+++ Legacy/src/gameObjects/card/hand-card.ts
@@ -1,8 +1,8 @@
-import { CARD_ASSETS_KEYS, CardAssetKeys } from '../../assets/asset-keys'
+import { CARD_ASSETS_KEYS, CardAssetKeys } from '../../../../src/assets/asset-keys'
 import { setOutline } from '../../common/outline'
 import { BattleScene } from '../../scenes/battle-scene'
-import { Coordinate } from '../../types/typedef'
-import { BATTLE_STATES, TARGET_KEYS, TargetKeys, WARNING_KEYS } from '../../utils/keys'
+import { Coordinate } from '../../../../src/types/typedef'
+import { BATTLE_STATES, TARGET_KEYS, TargetKeys, WARNING_KEYS } from '../../../../src/utils/keys'
 import { CARD_SCALE, PLAYER_BOARD_BOUNDS } from '../../utils/visual-configs'
 import { Card } from './card'
 import { CardData } from './card-keys'
