+++ Legacy/src/utils/battle-manager.ts
@@ -1,9 +1,9 @@
-import { EFFECT_ASSET_KEYS, UI_ASSET_KEYS } from '../assets/asset-keys'
+import { EFFECT_ASSET_KEYS, UI_ASSET_KEYS } from '../../../src/assets/asset-keys'
 import { Board } from '../gameObjects/board'
 import { BoardCard } from '../gameObjects/card/board-card'
 import { Hero } from '../gameObjects/hero'
 import { BattleScene } from '../scenes/battle-scene'
-import { BATTLE_STATES, TARGET_KEYS, TargetKeys } from './keys'
+import { BATTLE_STATES, TARGET_KEYS, TargetKeys } from '../../../src/utils/keys'
 import { ATTACK_CONFIGS, DAMAGE_CONFIGS, DEATH_CONFIGS } from './visual-configs'
 
 export class BattleManager {
