+++ src/assets/asset-keys.ts
